% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{abort_document_version_upload}
\alias{abort_document_version_upload}
\title{Aborts the upload of the specified document version that was previously initiated by InitiateDocumentVersionUpload}
\usage{
abort_document_version_upload(
  AuthenticationToken = "string",
  DocumentId = "string",
  VersionId = "string"
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The ID of the version.}
}
\description{
Aborts the upload of the specified document version that was previously initiated by InitiateDocumentVersionUpload. The client should make this call only when it no longer intends to upload the document version, or fails to do so.
}
\examples{

}
