% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{create_comment}
\alias{create_comment}
\title{Adds a new comment to the specified document version}
\usage{
create_comment(
  AuthenticationToken = "string",
  DocumentId = "string",
  VersionId = "string",
  ParentId = "string",
  ThreadId = "string",
  Text = "string",
  Visibility = "PUBLIC"|"PRIVATE",
  NotifyCollaborators = TRUE|FALSE
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The ID of the document version.}

\item{ParentId}{The ID of the parent comment.}

\item{ThreadId}{The ID of the root comment in the thread.}

\item{Text}{[required] The text of the comment.}

\item{Visibility}{The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.}

\item{NotifyCollaborators}{Set this parameter to TRUE to send an email out to the document collaborators after the comment is created.}
}
\description{
Adds a new comment to the specified document version.
}
\examples{

}
