% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{create_custom_metadata}
\alias{create_custom_metadata}
\title{Adds one or more custom properties to the specified resource (a folder, document, or version)}
\usage{
create_custom_metadata(ResourceId, CustomMetadata, AuthenticationToken = NULL,
  VersionId = NULL)
}
\arguments{
\item{ResourceId}{The ID of the resource.}

\item{CustomMetadata}{Custom metadata in the form of name-value pairs.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{VersionId}{The ID of the version, if the custom metadata is being added to a document version.}
}
\description{
Adds one or more custom properties to the specified resource (a folder, document, or version).
}
\examples{

}
