% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{create_folder}
\alias{create_folder}
\title{Creates a folder with the specified name and parent folder}
\usage{
create_folder(AuthenticationToken = NULL, Name = NULL, ParentFolderId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{Name}{The name of the new folder.}

\item{ParentFolderId}{[required] The ID of the parent folder.}
}
\description{
Creates a folder with the specified name and parent folder.
}
\section{Accepted Parameters}{
\preformatted{create_folder(
  AuthenticationToken = "string",
  Name = "string",
  ParentFolderId = "string"
)
}
}

