% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{create_labels}
\alias{create_labels}
\title{Adds the specified list of labels to the given resource (a document or folder)}
\usage{
create_labels(
  ResourceId = "string",
  Labels = list(
    "string"
  ),
  AuthenticationToken = "string"
)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource.}

\item{Labels}{[required] List of labels to add to the resource.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}
}
\description{
Adds the specified list of labels to the given resource (a document or folder)
}
\examples{

}
