% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{create_notification_subscription}
\alias{create_notification_subscription}
\title{Configure WorkDocs to use Amazon SNS notifications}
\usage{
create_notification_subscription(OrganizationId, Endpoint, Protocol,
  SubscriptionType)
}
\arguments{
\item{OrganizationId}{The ID of the organization.}

\item{Endpoint}{The endpoint to receive the notifications. If the protocol is HTTPS, the endpoint is a URL that begins with "https://".}

\item{Protocol}{The protocol to use. The supported value is https, which delivers JSON-encoded messages using HTTPS POST.}

\item{SubscriptionType}{The notification type.}
}
\description{
Configure WorkDocs to use Amazon SNS notifications.
}
\details{
The endpoint receives a confirmation message, and must confirm the subscription. For more information, see \href{http://docs.aws.amazon.com/sns/latest/dg/SendMessageToHttp.html#SendMessageToHttp.confirm}{Confirm the Subscription} in the \emph{Amazon Simple Notification Service Developer Guide}.
}
\examples{

}
