% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{create_user}
\alias{create_user}
\title{Creates a user in a Simple AD or Microsoft AD directory}
\usage{
create_user(OrganizationId = NULL, Username, EmailAddress = NULL,
  GivenName, Surname, Password, TimeZoneId = NULL, StorageRule = NULL,
  AuthenticationToken = NULL)
}
\arguments{
\item{OrganizationId}{The ID of the organization.}

\item{Username}{[required] The login name of the user.}

\item{EmailAddress}{The email address of the user.}

\item{GivenName}{[required] The given name of the user.}

\item{Surname}{[required] The surname of the user.}

\item{Password}{[required] The password of the user.}

\item{TimeZoneId}{The time zone ID of the user.}

\item{StorageRule}{The amount of storage for the user.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}
}
\description{
Creates a user in a Simple AD or Microsoft AD directory. The status of a newly created user is "ACTIVE". New users can access Amazon WorkDocs.
}
\section{Accepted Parameters}{
\preformatted{create_user(
  OrganizationId = "string",
  Username = "string",
  EmailAddress = "string",
  GivenName = "string",
  Surname = "string",
  Password = "string",
  TimeZoneId = "string",
  StorageRule = list(
    StorageAllocatedInBytes = 123,
    StorageType = "UNLIMITED"|"QUOTA"
  ),
  AuthenticationToken = "string"
)
}
}

