% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{deactivate_user}
\alias{deactivate_user}
\title{Deactivates the specified user, which revokes the user's access to Amazon WorkDocs}
\usage{
deactivate_user(UserId, AuthenticationToken = NULL)
}
\arguments{
\item{UserId}{[required] The ID of the user.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}
}
\description{
Deactivates the specified user, which revokes the user's access to Amazon WorkDocs.
}
\section{Accepted Parameters}{
\preformatted{deactivate_user(
  UserId = "string",
  AuthenticationToken = "string"
)
}
}

