% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{delete_comment}
\alias{delete_comment}
\title{Deletes the specified comment from the document version}
\usage{
delete_comment(DocumentId, VersionId, CommentId, AuthenticationToken = NULL)
}
\arguments{
\item{DocumentId}{The ID of the document.}

\item{VersionId}{The ID of the document version.}

\item{CommentId}{The ID of the comment.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}
}
\description{
Deletes the specified comment from the document version.
}
\examples{

}
