% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{delete_document}
\alias{delete_document}
\title{Permanently deletes the specified document and its associated metadata}
\usage{
delete_document(DocumentId, AuthenticationToken = NULL)
}
\arguments{
\item{DocumentId}{The ID of the document.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}
}
\description{
Permanently deletes the specified document and its associated metadata.
}
\examples{

}
