% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{delete_folder}
\alias{delete_folder}
\title{Permanently deletes the specified folder and its contents}
\usage{
delete_folder(FolderId, AuthenticationToken = NULL)
}
\arguments{
\item{FolderId}{The ID of the folder.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}
}
\description{
Permanently deletes the specified folder and its contents.
}
\examples{

}
