% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{delete_folder_contents}
\alias{delete_folder_contents}
\title{Deletes the contents of the specified folder}
\usage{
delete_folder_contents(FolderId, AuthenticationToken = NULL)
}
\arguments{
\item{FolderId}{The ID of the folder.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}
}
\description{
Deletes the contents of the specified folder.
}
\examples{

}
