% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{delete_labels}
\alias{delete_labels}
\title{Deletes the specified list of labels from a resource}
\usage{
delete_labels(
  ResourceId = "string",
  AuthenticationToken = "string",
  Labels = list(
    "string"
  ),
  DeleteAll = TRUE|FALSE
)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{Labels}{List of labels to delete from the resource.}

\item{DeleteAll}{Flag to request removal of all labels from the specified resource.}
}
\description{
Deletes the specified list of labels from a resource.
}
\examples{

}
