% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{delete_notification_subscription}
\alias{delete_notification_subscription}
\title{Deletes the specified subscription from the specified organization}
\usage{
delete_notification_subscription(SubscriptionId, OrganizationId)
}
\arguments{
\item{SubscriptionId}{The ID of the subscription.}

\item{OrganizationId}{The ID of the organization.}
}
\description{
Deletes the specified subscription from the specified organization.
}
\examples{

}
