% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{delete_user}
\alias{delete_user}
\title{Deletes the specified user from a Simple AD or Microsoft AD directory}
\usage{
delete_user(
  AuthenticationToken = "string",
  UserId = "string"
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{UserId}{[required] The ID of the user.}
}
\description{
Deletes the specified user from a Simple AD or Microsoft AD directory.
}
\examples{

}
