% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{describe_comments}
\alias{describe_comments}
\title{List all the comments for the specified document version}
\usage{
describe_comments(
  AuthenticationToken = "string",
  DocumentId = "string",
  VersionId = "string",
  Limit = 123,
  Marker = "string"
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The ID of the document version.}

\item{Limit}{The maximum number of items to return.}

\item{Marker}{The marker for the next set of results. This marker was received from a previous call.}
}
\description{
List all the comments for the specified document version.
}
\examples{

}
