% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{describe_root_folders}
\alias{describe_root_folders}
\title{Describes the current user's special folders; the RootFolder and the RecycleBin}
\usage{
describe_root_folders(AuthenticationToken, Limit = NULL, Marker = NULL)
}
\arguments{
\item{AuthenticationToken}{[required] Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{Limit}{The maximum number of items to return.}

\item{Marker}{The marker for the next set of results. (You received this marker from a previous call.)}
}
\description{
Describes the current user's special folders; the \code{RootFolder} and the \code{RecycleBin}. \code{RootFolder} is the root of user's files and folders and \code{RecycleBin} is the root of recycled items. This is not a valid action for SigV4 (administrative API) clients.
}
\section{Accepted Parameters}{
\preformatted{describe_root_folders(
  AuthenticationToken = "string",
  Limit = 123,
  Marker = "string"
)
}
}

