% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{get_document}
\alias{get_document}
\title{Retrieves details of a document}
\usage{
get_document(AuthenticationToken = NULL, DocumentId,
  IncludeCustomMetadata = NULL)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{DocumentId}{[required] The ID of the document.}

\item{IncludeCustomMetadata}{Set this to \code{TRUE} to include custom metadata in the response.}
}
\description{
Retrieves details of a document.
}
\section{Accepted Parameters}{
\preformatted{get_document(
  AuthenticationToken = "string",
  DocumentId = "string",
  IncludeCustomMetadata = TRUE|FALSE
)
}
}

