% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{get_folder}
\alias{get_folder}
\title{Retrieves the metadata of the specified folder}
\usage{
get_folder(AuthenticationToken = NULL, FolderId,
  IncludeCustomMetadata = NULL)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{FolderId}{[required] The ID of the folder.}

\item{IncludeCustomMetadata}{Set to TRUE to include custom metadata in the response.}
}
\description{
Retrieves the metadata of the specified folder.
}
\section{Accepted Parameters}{
\preformatted{get_folder(
  AuthenticationToken = "string",
  FolderId = "string",
  IncludeCustomMetadata = TRUE|FALSE
)
}
}

