% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{get_folder_path}
\alias{get_folder_path}
\title{Retrieves the path information (the hierarchy from the root folder) for the specified folder}
\usage{
get_folder_path(FolderId, AuthenticationToken = NULL, Limit = NULL,
  Fields = NULL, Marker = NULL)
}
\arguments{
\item{FolderId}{The ID of the folder.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{Limit}{The maximum number of levels in the hierarchy to return.}

\item{Fields}{A comma-separated list of values. Specify "NAME" to include the names of the parent folders.}

\item{Marker}{This value is not supported.}
}
\description{
Retrieves the path information (the hierarchy from the root folder) for the specified folder.
}
\details{
By default, Amazon WorkDocs returns a maximum of 100 levels upwards from the requested folder and only includes the IDs of the parent folders in the path. You can limit the maximum number of levels. You can also request the parent folder names.
}
\examples{

}
