% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{get_resources}
\alias{get_resources}
\title{Retrieves a collection of resources, including folders and documents}
\usage{
get_resources(AuthenticationToken = NULL, UserId = NULL,
  CollectionType = NULL, Limit = NULL, Marker = NULL)
}
\arguments{
\item{AuthenticationToken}{The Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API operation using AWS credentials.}

\item{UserId}{The user ID for the resource collection. This is a required field for accessing the API operation using IAM credentials.}

\item{CollectionType}{The collection type.}

\item{Limit}{The maximum number of resources to return.}

\item{Marker}{The marker for the next set of results. This marker was received from a previous call.}
}
\description{
Retrieves a collection of resources, including folders and documents. The only \code{CollectionType} supported is \code{SHARED_WITH_ME}.
}
\section{Accepted Parameters}{
\preformatted{get_resources(
  AuthenticationToken = "string",
  UserId = "string",
  CollectionType = "SHARED_WITH_ME",
  Limit = 123,
  Marker = "string"
)
}
}

