% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{initiate_document_version_upload}
\alias{initiate_document_version_upload}
\title{Creates a new document object and version object}
\usage{
initiate_document_version_upload(ParentFolderId, AuthenticationToken = NULL,
  Id = NULL, Name = NULL, ContentCreatedTimestamp = NULL,
  ContentModifiedTimestamp = NULL, ContentType = NULL,
  DocumentSizeInBytes = NULL)
}
\arguments{
\item{ParentFolderId}{The ID of the parent folder.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{Id}{The ID of the document.}

\item{Name}{The name of the document.}

\item{ContentCreatedTimestamp}{The timestamp when the content of the document was originally created.}

\item{ContentModifiedTimestamp}{The timestamp when the content of the document was modified.}

\item{ContentType}{The content type of the document.}

\item{DocumentSizeInBytes}{The size of the document, in bytes.}
}
\description{
Creates a new document object and version object.
}
\details{
The client specifies the parent folder ID and name of the document to upload. The ID is optionally specified when creating a new version of an existing document. This is the first step to upload a document. Next, upload the document to the URL returned from the call, and then call UpdateDocumentVersion.

To cancel the document upload, call AbortDocumentVersionUpload.
}
\examples{

}
