% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_package.R
\docType{package}
\name{paws.workdocs-package}
\alias{paws.workdocs}
\alias{paws.workdocs-package}
\title{paws.workdocs: Amazon WorkDocs}
\description{
- File Migration: File migration applications are supported
for users who want to migrate their files from an on-premises or
off-premises file system or service. Users can insert files into a
user directory structure, as well as allow for basic metadata changes,
such as modifications to the permissions of files.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
