% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{remove_all_resource_permissions}
\alias{remove_all_resource_permissions}
\title{Removes all the permissions from the specified resource}
\usage{
remove_all_resource_permissions(ResourceId, AuthenticationToken = NULL)
}
\arguments{
\item{ResourceId}{The ID of the resource.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}
}
\description{
Removes all the permissions from the specified resource.
}
\examples{

}
