% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{update_document_version}
\alias{update_document_version}
\title{Changes the status of the document version to ACTIVE}
\usage{
update_document_version(AuthenticationToken = NULL, DocumentId,
  VersionId, VersionStatus = NULL)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The version ID of the document.}

\item{VersionStatus}{The status of the version.}
}
\description{
Changes the status of the document version to ACTIVE.
}
\details{
Amazon WorkDocs also sets its document container to ACTIVE. This is the last step in a document upload, after the client uploads the document to an S3-presigned URL returned by InitiateDocumentVersionUpload.
}
\section{Accepted Parameters}{
\preformatted{update_document_version(
  AuthenticationToken = "string",
  DocumentId = "string",
  VersionId = "string",
  VersionStatus = "ACTIVE"
)
}
}

