% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{add_resource_permissions}
\alias{add_resource_permissions}
\title{Creates a set of permissions for the specified folder or document}
\usage{
add_resource_permissions(
  AuthenticationToken = "string",
  ResourceId = "string",
  Principals = list(
    list(
      Id = "string",
      Type = "USER"|"GROUP"|"INVITE"|"ANONYMOUS"|"ORGANIZATION",
      Role = "VIEWER"|"CONTRIBUTOR"|"OWNER"|"COOWNER"
    )
  ),
  NotificationOptions = list(
    SendEmail = TRUE|FALSE,
    EmailMessage = "string"
  )
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{ResourceId}{[required] The ID of the resource.}

\item{Principals}{[required] The users, groups, or organization being granted permission.}

\item{NotificationOptions}{The notification options.}
}
\description{
Creates a set of permissions for the specified folder or document. The resource permissions are overwritten if the principals already have different permissions.
}
\examples{

}
