% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{create_user}
\alias{create_user}
\title{Creates a user in a Simple AD or Microsoft AD directory}
\usage{
create_user(Username, GivenName, Surname, Password, OrganizationId = NULL,
  EmailAddress = NULL, TimeZoneId = NULL, StorageRule = NULL,
  AuthenticationToken = NULL)
}
\arguments{
\item{Username}{The login name of the user.}

\item{GivenName}{The given name of the user.}

\item{Surname}{The surname of the user.}

\item{Password}{The password of the user.}

\item{OrganizationId}{The ID of the organization.}

\item{EmailAddress}{The email address of the user.}

\item{TimeZoneId}{The time zone ID of the user.}

\item{StorageRule}{The amount of storage for the user.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}
}
\description{
Creates a user in a Simple AD or Microsoft AD directory. The status of a newly created user is "ACTIVE". New users can access Amazon WorkDocs.
}
\examples{

}
