% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{delete_custom_metadata}
\alias{delete_custom_metadata}
\title{Deletes custom metadata from the specified resource}
\usage{
delete_custom_metadata(
  AuthenticationToken = "string",
  ResourceId = "string",
  VersionId = "string",
  Keys = list(
    "string"
  ),
  DeleteAll = TRUE|FALSE
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{ResourceId}{[required] The ID of the resource, either a document or folder.}

\item{VersionId}{The ID of the version, if the custom metadata is being deleted from a document version.}

\item{Keys}{List of properties to remove.}

\item{DeleteAll}{Flag to indicate removal of all custom metadata properties from the specified resource.}
}
\description{
Deletes custom metadata from the specified resource.
}
\examples{

}
