% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{delete_document}
\alias{delete_document}
\title{Permanently deletes the specified document and its associated metadata}
\usage{
delete_document(
  AuthenticationToken = "string",
  DocumentId = "string"
)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{DocumentId}{[required] The ID of the document.}
}
\description{
Permanently deletes the specified document and its associated metadata.
}
\examples{

}
