% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{describe_groups}
\alias{describe_groups}
\title{Describes the groups specified by query}
\usage{
describe_groups(AuthenticationToken = NULL, SearchQuery,
  OrganizationId = NULL, Marker = NULL, Limit = NULL)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{SearchQuery}{[required] A query to describe groups by group name.}

\item{OrganizationId}{The ID of the organization.}

\item{Marker}{The marker for the next set of results. (You received this marker from a previous call.)}

\item{Limit}{The maximum number of items to return with this call.}
}
\description{
Describes the groups specified by query.
}
\section{Accepted Parameters}{
\preformatted{describe_groups(
  AuthenticationToken = "string",
  SearchQuery = "string",
  OrganizationId = "string",
  Marker = "string",
  Limit = 123
)
}
}

