% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{describe_resource_permissions}
\alias{describe_resource_permissions}
\title{Describes the permissions of a specified resource}
\usage{
describe_resource_permissions(ResourceId, AuthenticationToken = NULL,
  PrincipalId = NULL, Limit = NULL, Marker = NULL)
}
\arguments{
\item{ResourceId}{The ID of the resource.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{PrincipalId}{The ID of the principal to filter permissions by.}

\item{Limit}{The maximum number of items to return with this call.}

\item{Marker}{The marker for the next set of results. (You received this marker from a previous call)}
}
\description{
Describes the permissions of a specified resource.
}
\examples{

}
