% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{get_document_version}
\alias{get_document_version}
\title{Retrieves version metadata for the specified document}
\usage{
get_document_version(AuthenticationToken = NULL, DocumentId, VersionId,
  Fields = NULL, IncludeCustomMetadata = NULL)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{DocumentId}{[required] The ID of the document.}

\item{VersionId}{[required] The version ID of the document.}

\item{Fields}{A comma-separated list of values. Specify "SOURCE" to include a URL for the source document.}

\item{IncludeCustomMetadata}{Set this to TRUE to include custom metadata in the response.}
}
\description{
Retrieves version metadata for the specified document.
}
\section{Accepted Parameters}{
\preformatted{get_document_version(
  AuthenticationToken = "string",
  DocumentId = "string",
  VersionId = "string",
  Fields = "string",
  IncludeCustomMetadata = TRUE|FALSE
)
}
}

