% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{remove_resource_permission}
\alias{remove_resource_permission}
\title{Removes the permission for the specified principal from the specified resource}
\usage{
remove_resource_permission(ResourceId, PrincipalId,
  AuthenticationToken = NULL, PrincipalType = NULL)
}
\arguments{
\item{ResourceId}{The ID of the resource.}

\item{PrincipalId}{The principal ID of the resource.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{PrincipalType}{The principal type of the resource.}
}
\description{
Removes the permission for the specified principal from the specified resource.
}
\examples{

}
