% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workdocs_operations.R
\name{update_folder}
\alias{update_folder}
\title{Updates the specified attributes of the specified folder}
\usage{
update_folder(FolderId, AuthenticationToken = NULL, Name = NULL,
  ParentFolderId = NULL, ResourceState = NULL)
}
\arguments{
\item{FolderId}{The ID of the folder.}

\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.}

\item{Name}{The name of the folder.}

\item{ParentFolderId}{The ID of the parent folder.}

\item{ResourceState}{The resource state of the folder. Only ACTIVE and RECYCLED are accepted values from the API.}
}
\description{
Updates the specified attributes of the specified folder. The user must have access to both the folder and its parent folder, if applicable.
}
\examples{

}
